+++ deployables/extension/src/panel/pages/$activeRouteId/ActiveRoute.tsx
@@ -5,11 +5,6 @@ import {
   saveLastUsedRouteId,
 } from '@/execution-routes'
 import { ProvideProvider } from '@/providers-ui'
-import { getActiveTab, sendMessageToTab } from '@/utils'
-import {
-  CompanionResponseMessageType,
-  type CompanionResponseMessage,
-} from '@zodiac/messages'
 import {
   Outlet,
   redirect,
@@ -26,19 +21,6 @@ export const loader = async ({ params }: LoaderFunctionArgs) => {
     const route = await getRoute(activeRouteId)
 
     await saveStorageEntry({ key: 'lastUsedRoute', value: route.id })
-
-    const activeTab = await getActiveTab()
-
-    if (activeTab.id != null) {
-      sendMessageToTab(
-        activeTab.id,
-        {
-          type: CompanionResponseMessageType.PROVIDE_ACTIVE_ROUTE,
-          activeRouteId: route.id,
-        } satisfies CompanionResponseMessage,
-        { protocolCheckOnly: true },
-      )
-    }
 
     return { route: await markRouteAsUsed(route) }
   } catch {
