+++ deployables/extension/src/panel/pages/Root.tsx
@@ -1,6 +1,6 @@
 import { ProvideCompanionAppContext } from '@/companion'
 import { getLastUsedRouteId } from '@/execution-routes'
-import { sendMessageToTab } from '@/utils'
+import { getActiveTab, sendMessageToTab } from '@/utils'
 import { getCompanionAppUrl } from '@zodiac/env'
 import {
   CompanionAppMessageType,
@@ -16,6 +16,18 @@ import { useSaveRoute } from './useSaveRoute'
 
 export const loader = async () => {
   const lastUsedRouteId = await getLastUsedRouteId()
+
+  const activeTab = await getActiveTab()
+  if (activeTab.id != null) {
+    sendMessageToTab(
+      activeTab.id,
+      {
+        type: CompanionResponseMessageType.PROVIDE_ACTIVE_ROUTE,
+        activeRouteId: lastUsedRouteId,
+      } satisfies CompanionResponseMessage,
+      { protocolCheckOnly: true },
+    )
+  }
 
   return { lastUsedRouteId, companionAppUrl: getCompanionAppUrl() }
 }
